devtools/client/shared/components/reps/test/mochitest/test_reps_string.html

<!-- This Source Code Form is subject to the terms of the Mozilla Public
   - License, v. 2.0. If a copy of the MPL was not distributed with this
   - file, You can obtain one at http://mozilla.org/MPL/2.0/. -->
<!DOCTYPE HTML>
<html>
<!--
Test String rep
-->
<head>
  <meta charset="utf-8">
  <title>Rep test - String</title>
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css">
</head>
<body>
<pre id="test">
<script src="head.js" type="application/javascript"></script>
<script type="application/javascript">
window.onload = Task.async(function* () {
  const {
    REPS,
    getRep,
   } = browserRequire("devtools/client/shared/components/reps/reps");
  let { StringRep } = REPS;

  const test_cases = [{
    name: "testMultiline",
    props: {
      object: "aaaaaaaaaaaaaaaaaaaaa\nbbbbbbbbbbbbbbbbbbb\ncccccccccccccccc\n",
    },
    result: "\"aaaaaaaaaaaaaaaaaaaaa\\nbbbbbbbbbbbbbbbbbbb\\ncccccccccccccccc\\n\""
  }, {
    name: "testMultilineLimit",
    props: {
      object: "aaaaaaaaaaaaaaaaaaaaa\nbbbbbbbbbbbbbbbbbbb\ncccccccccccccccc\n",
      cropLimit: 20
    },
    result: "\"aaaaaaaaa…cccccc\\n\""
  }, {
    name: "testMultilineOpen",
    props: {
      object: "aaaaaaaaaaaaaaaaaaaaa\nbbbbbbbbbbbbbbbbbbb\ncccccccccccccccc\n",
      member: { open: true }
    },
    result: "\"aaaaaaaaaaaaaaaaaaaaa\\nbbbbbbbbbbbbbbbbbbb\\ncccccccccccccccc\\n\""
  }, {
    name: "testUseQuotes",
    props: {
      object: "abc",
      useQuotes: false
    },
    result: "abc"
  }, {
    name: "testNonPrintableCharacters",
    props: {
      object: "a\x01b",
      useQuotes: false
    },
    result: "a\ufffdb"
  }, {
    name: "testQuoting",
    props: {
      object: "\t\n\r\"'\\\x1f\x9f\ufeff\ufffe\ud8000\u2063\ufffc\u2028\ueeee",
      useQuotes: true
    },
    result: "\"\\t\\n\\r\\\"'\\\\\\u001f\\u009f\\ufeff\\ufffe\\ud8000\\u2063\\ufffc\\u2028\\ueeee\""
  }, {
    name: "testUnpairedSurrogate",
    props: {
      object: "\uDC23",
      useQuotes: true
    },
    result: "\"\\udc23\""
  }, {
    name: "testValidSurrogate",
    props: {
      object: "\ud83d\udeec",
      useQuotes: true
    },
    result: "\"\ud83d\udeec\""
  }, {
    name: "testNoEscapeWhitespace",
    props: {
      object: "line 1\r\nline 2\n\tline 3",
      useQuotes: true,
      escapeWhitespace: false,
    },
    result: "\"line 1\r\nline 2\n\tline 3\""
  }];

  try {
    // Test that correct rep is chosen
    is(getRep(test_cases[0].props.object), StringRep.rep,
      "Rep correctly selects String Rep");

    // Test rendering
    for (let test of test_cases) {
      const renderedComponent = renderComponent(StringRep.rep, test.props);
      is(renderedComponent.textContent, test.result, "String rep " + test.name);
    }
  } catch(e) {
    ok(false, "Got an error: " + DevToolsUtils.safeErrorString(e));
  } finally {
    SimpleTest.finish();
  }
});
</script>
</pre>
</body>
</html>
